Organize files in different folders
Add myTray class that extends Tray.
