+++ app/assets/images/icon.png
