+++ app/css/index.css
@@ -13,7 +13,7 @@ body{
 main{
     height: 100%;
     width: 100%;
-    background-image: url('./assets/image/1.jpg');
+    background-image: url('../assets/images/1.jpg');
     display: flex;
     flex-direction: column;
     justify-content: space-between;
