+++ app/renderer.js
