+++ package.json
@@ -2,7 +2,7 @@
     "name": "electron101",
     "version": "1.0.0",
     "description": "",
-    "main": "index.js",
+    "main": "app/index.js",
     "scripts": {
         "start": "electronmon ."
     },
